Refactor tournament seeding module for improved structure
Restructure and enhance the tournament seeding module in the BowlingMegabucks application. Introduced new interfaces and classes (`IView`, `IAdapter`, `IViewModel`) to define behavior and data structures. Updated the `Presenter` class for better presentation logic, including error handling and data binding. Modified `TournamentsExtensions` to register new seeding services with the dependency injection container, improving modularity, readability, and maintainability.
