--- source/main/Tournaments/Seeding/ITournamentSeedingView.cs
@@ -1,10 +1,0 @@
-﻿
-namespace BowlingMegabucks.TournamentManager.Tournaments.Seeding;
-internal interface IView
-{
-    TournamentId Id { get; }
-
-    void DisplayError(string message);
-
-    void BindResults(string divisionName, ICollection<IViewModel> scores);
-}
