+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Tournaments/Seeding/TournamentSeedingAdapter.cs
